grfs small bugfixes
Reading: https://sites.google.com/site/jdquirk/articles/eco-extremism

I am a bit of an ecoextremist myself. And I too have grown wary of the
end of the world scenarios promoted by some environmentalists. I for
one don't find the idea of global warming and the collapse of civiliztaion
to be terribly frightening, and I do not believe that other eco-extremists
do either.

Some of those eco-extremists, who hawk global warming catastrophe, wrongly
believe or hope, that there exists some fundamental moral balance to this
world. A balance which will mean that the boring suburban uper middleclass
will be automatically punished for the tremendous ecological damage that
their comfortable lives is causing.

This belief is wrongheaded on two fronts. On the factual front, the world
does not tend towards moral balance. Without moral intervention, life
tends towards being extremely unfair. The hope that those who live ecologically
unsound lives will be punished for doing so is simply a wild fantasy. On the
moral front, the catastrophic line of argument is morally perverted. Telling someone
that they need to start living green or they will die in a horific global
warming induced fireball is not moral argument, it is barganing. While I have
nothing against barganing with those who don't give a shit about the environment,
we should understand that we are sitting down to a business transaction when we
do so. Environmentalists would have far more barganing power if they were to use
immediate threats, such as the nationalistic fear of dependence on other countries
natural resources, rather than abstract ones.

Mr. Quirk writes "but even the simultaneous detonation of every thermonuclear
device stockpiled on the face of the Earth would result in only a minor shudder
to Mother Nature - a thousand years later, Earth would look largely the same as
it does now, except without humans and a few other species." And yet this is
exactly what some neo-primitivist thinkers, such as Ted Kaczinski, want. The
claim that nature can heal itself when left to its own devices is hardly a valid
rebuttle of a philosophy which wants just that.

I personally do not believe that there will be a catastrophe which will shake
the comfort of the suburban middle class. We can always drill deeper, and while
old growth lumber may cease to exist, lime stone is harldy running out. They
are well capable of covering the entire world with soybeans, concrete, and flat
boring lawns for the next ten thousand years. And living out their meaningless lives
in peace. But I don't want that. Indeed, that is exactly what I fear most.

---

Many comfortable suburbanites are happy. I am not. How can I convince them to
change their ways? I can try to make them feel bad by telling them about the
problems in todays world. But they are isolated from such problems far too well.
I could threaten them with violence or economic sanctions. But the truth is that
I am neither a military nor an economic power. Even moral arguments seem to
be made of sand. You tell a suburbanite that the suffering of animals in factory
farms is bad and they will concure. Countering that nature contains fundamental
suffering. The dear must either be eaten by the wolf or die from over-population,
desease and starvation. Nature is bad, the suburbanite can truthfully counter.
So if there is a moral argument, perhaps it is for the destruction of nature.

I can assert myself on esthetic grounds. After-all nature *is* beautiful. But
aethetics has two problems. For one, it is subjective. More woryingly, aesthetics
is the domain of right wing extremists. There is no subjective dissagreement between
Mr. Goebbels and I about the beauty of his children. I do find them beautiful.
If I am to argue that we should fight for nature on account of it's beauty, then
should we fight for the Aryan race as well?

The truth is, I haven't got any good argument. I only have the experience of
a lonely childhood in upper middle class suburbia. The view of 7 acre parking
lots. The smell of hot asphalt. The hatred of the bordom and the meaninglessness.
The love of the tree trunk. Subjective things I know.

Listening to: https://www.youtube.com/watch?v=doS3hYIsFt4